more poison added

+++ common/include/mm/new.h
@@ -1,8 +1,11 @@
 //----------------------------------------------------------------------
-//   $Id: new.h,v 1.2 2005/04/22 17:21:40 nomenquis Exp $
+//   $Id: new.h,v 1.3 2005/04/22 19:43:04 nomenquis Exp $
 //----------------------------------------------------------------------
 //
 //  $Log: new.h,v $
+//  Revision 1.2  2005/04/22 17:21:40  nomenquis
+//  added TONS of stuff, changed ZILLIONS of things
+//
 //  Revision 1.1  2005/04/20 21:35:33  nomenquis
 //  started to implement page manager
 //
@@ -23,8 +26,8 @@ inline void* operator new(size_t, void* p)
 }
 
 void __builtin_delete(void* address);
-void* __builtin_new(unsigned long size);
-void* __builtin_vec_new(unsigned long size);
+void* __builtin_new(uint32 size);
+void* __builtin_vec_new(uint32 size);
 void __builtin_vec_delete(void* address);
